+++ apps/backend/src/libs/modules/server-application/server-application.ts
@@ -1,6 +1,7 @@
 import { config } from "~/libs/modules/config/config.js";
 import { database } from "~/libs/modules/database/database.js";
 import { logger } from "~/libs/modules/logger/logger.js";
+import { activityLogController } from "~/modules/activity-logs/activity-logs.js";
 import { authController } from "~/modules/auth/auth.js";
 import { groupController } from "~/modules/groups/groups.js";
 import { permissionController } from "~/modules/permissions/permissions.js";
@@ -18,6 +19,7 @@ import { WHITE_ROUTES } from "./libs/constants/constants.js";
 const apiV1 = new BaseServerApplicationApi(
 	"v1",
 	config,
+	...activityLogController.routes,
 	...authController.routes,
 	...permissionController.routes,
 	...projectGroupController.routes,