+++ apps/backend/src/modules/project-groups/project-group.repository.ts
@@ -52,7 +52,7 @@ class ProjectGroupRepository implements Repository {
 			await trx.rollback();
 
 			throw new ProjectGroupError({
-				message: ExceptionMessage.CREATE_PROJECT_GROUP_FAILED,
+				message: ExceptionMessage.PROJECT_GROUP_CREATE_FAILED,
 				status: HTTPCode.INTERNAL_SERVER_ERROR,
 			});
 		}