+++ apps/backend/src/modules/users/libs/types/types.ts
@@ -1,6 +1,7 @@
 export {
 	type UserAuthResponseDto,
 	type UserGetAllItemResponseDto,
+	type UserGetAllQueryParameters,
 	type UserGetAllResponseDto,
 	type UserPatchRequestDto,
 	type UserPatchResponseDto,