+++ apps/frontend/src/pages/projects/libs/components/project-search/project-search.tsx
@@ -9,7 +9,12 @@ type Properties = {
 const ProjectsSearch = ({ onChange }: Properties): JSX.Element => {
 	return (
 		<div className={styles["search-container"]}>
-			<Search isLabelHidden label="Projects search" onChange={onChange} />
+			<Search
+				isLabelHidden
+				label="Projects search"
+				onChange={onChange}
+				placeholder="Enter project name"
+			/>
 		</div>
 	);
 };